split DVGeometry; using baseclasses Error
